% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.hifld.urgent-package.R
\docType{package}
\name{dc.hifld.urgent-package}
\alias{dc.hifld.urgent}
\alias{dc.hifld.urgent-package}
\title{dc.hifld.urgent: Social Data Commons: Health, Behavior, Diet: Urgent Care Locations}
\description{
Allows user to easily get urgent care locations data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Eric Oh \email{eo9tv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
